% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newhalfviolinplot.R
\name{myplot}
\alias{myplot}
\title{new half violin plot with 2 conditions}
\usage{
myplot(
  df,
  x_var,
  y_var,
  fill_var,
  condition1,
  condition2,
  color1,
  color2,
  topic,
  number
)
}
\arguments{
\item{df}{your own imported data frame}

\item{x_var}{one of column names and it should be a character}

\item{y_var}{one of column names and it should be a character}

\item{fill_var}{one of column names containing only 2 values and it should be a character}

\item{condition1}{value 1 in fill_var and it should be a character}

\item{condition2}{value 2 in fill_var and it should be a character}

\item{color1}{the color you like and it should be a character}

\item{color2}{the color you like and it should be a character}

\item{topic}{the main content of your plot and it should be a character}

\item{number}{adjust the position}
}
\value{
p
}
\description{
new half violin plot with 2 conditions
}
\examples{
df<-read.csv("testdata.csv")
myplot(df,"celltype_final","ATP5MC3","condition","CTRL","EXP","pink","blue","myplot",0.05)
}
